Title: 이차원 배열의 오름차순 정리, Time: 124ms, Memory: 8MB, Status: Passed - CodeTree

+++ 241023/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.java
@@ -4,7 +4,7 @@ import java.util.*;
 public class Main {
     public static final int MAX_N = 100000;
 
-    public static int n, k;
+    public static long n, k;
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
